--- app/page.module.css
@@ -1,3 +1,0 @@
-.main-container{
-  
-}
